Separating components in 2 files
